add vectors update on points update
